+++ src/config.js
@@ -5,14 +5,13 @@ const globalConfig = reactive({
     size: 'medium',
     zIndex: 2000,
     theme: 'light'
-  })
-  
-  export const useGlobalConfig = () => {
+})
+
+export const useGlobalConfig = () => {
     return {
-      config: globalConfig,
-      setConfig: (key, value) => {
-        globalConfig[key] = value
-      }
+        config: globalConfig,
+        setConfig: (key, value) => {
+            globalConfig[key] = value
+        }
     }
-  }
-  
+}