+++ src/main.js
@@ -1,12 +1,13 @@
 import { createApp } from 'vue'
 // import MyComponents from './index.js';
 import App from './App.vue';
-import  install  from './index.js'
+import install from './index.js'
 
 const app = createApp(App);
-await install(app);
+(async() => {
+    await install(app);
+    app.mount('#app');
+})();
 
 // app.use(MyComponents)
-
-app.mount('#app')
 // createApp().use(MyComponents)